requirements.txt
absl-py==1.0.0
numpy==1.21.5
onnxruntime==1.10.0
opencv_python==4.5.5.62
